requirements.txt
git+https://github.com/Rapptz/discord.py
PyNaCl==1.3.0
pandas
python-dotenv
dnspython==1.16.0
async-timeout==3.0.1
discord-webhook
